pyproject.toml
[project]
name = "daily-note-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "google-genai>=1.65.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-frontmatter>=1.1.0",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
